talked to jack, no need to scale
this math is soooo clean. so basically, if your voltage reference is 4.096 and you have a 12 bit adc message, that comes to 0.001 mV and since we want volts thats just 1V so really, you don't have to scale anything and you get your voltage straight from the proper channels. waohhhhh