Log and manage power controls of the charging pad (#1)
* include necessary mcc files

contents copied from cansw_actuator repo

* Customize boiler plate

also added comments to map out what needs to be done

* Add voltage health can messages

* forgot multiplier for VSENSE

* im done for tonight

need to figure out which adcc_channels I need and also what constant to scael the getSingleConversion value by

* I found the adcc_channel values for my pins (:

* talked to jack, no need to scale

this math is soooo clean. so basically, if your voltage reference is 4.096 and you have a 12 bit adc message, that comes to 0.001 mV and since we want volts thats just 1V so really, you don't have to scale anything and you get your voltage straight from the proper channels. waohhhhh

* add remaining pin initialized values

* thanks to joe, i think this is 'done'

need to test the default values for 5V_line and BATT_on to see if its a 0 or 1, and then also create (?) sensor IDs for the 5 and 13volt lines

* add vscode to gitignore

* change nplab project setting name to cansw_charging

* configure mplab nbproject settings

mcc_genereated_files are now linked and I fixed up the file structure

* code should be good now

probably need to create a new board sensor id for the 5v and 13v current lines?

* Fixed voltage logging values

tldr; had to slow down the clock to sample the analog-to-digital-converter properly, only spent half a day (:

* I think i mixed up my tris values

tris=1 is input, aka you read in while tris=0 is output, so you write the output of the pin

* Board accurately responds to CAN messages

MSG_ACTUATOR_CMD ACTUATOR OPEN/CLOSED turns the 5V line on or off respectivelynow

* Update can_msg sensor ids

need to create sensor_charge curr and volt in canlib for the sensor_id

* Fix charging current leaking

basically I had the battery charger value set incorrectly (forgot to test)

* Cleanup comments

* turns out the sketchy adc channel stays

jack mentioned that creating the adcc channel enums in the adcc.h file was a bit sketchy, but since there are functions that reference the enum type, it needs to be defined there as well

* rename variables to be clearer

* Yoink cansw_arming low_pass_filter

Just need to verify this in person (can toggle low_pass_response to control the weightings and MAX_SENSOR_LOOP_TIME_DIFF_ms for how frequent to modify said average

* Tested low pass filter

there was also a mistake with the current scalar factor, was off by a factor of 2

* Update code

canlib changed, board respin happened, fixed some issues where I was multipling instead of dividing (need to confirm by testing it in person)

* address bugs

* Update canlib commit id

* WDR fixes

* Remove TODOs

if it aint broken, why fix it

---------

Co-authored-by: Michael Li <michael.li.web@gmail.co>
Co-authored-by: BluCodeGH <22122986+BluCodeGH@users.noreply.github.com>